Initial commit of a the rollingops framework.
This is an example charm showcasing an implementation of the Rolling
Operations library.

The library is included with the example charm.

This is a first take on an asynchronous lock built on top of the
primitives in the Operator Framework. It passes unit tests, but has
not been extensively tested in a live charm.