Namespaced events, and fixed up inheritance.
Refactored the RollingOpsManager so that it defines a series of
namespaced events. This allows us to auto attach them, without
worrying too much about clobbering.

Also simplified the way that we're passing Framework state around, to
rely more closely on the Object meta class.

Fixed statuses. We now set an application wide status, which the smoke
test watches for.